+++ AntJob.Data/Entity/应用消息.Biz.cs
@@ -50,7 +50,7 @@ namespace AntJob.Data.Entity
         /// <summary>根据编号查找</summary>
         /// <param name="id">编号</param>
         /// <returns>实体对象</returns>
-        public static AppMessage FindByID(Int32 id)
+        public static AppMessage FindByID(Int64 id)
         {
             if (id <= 0) return null;
